refactor(arch): unify plotting in epione.pl by chart type (PR 3/5)
Visualization is modality-agnostic; the v0.4 design holds all plotting
helpers in :mod:`epione.pl`, organised by chart type rather than by
modality.

This PR consolidates the Hi-C plot family:

  bulk/hic/_plot.py         → pl/_contact.py
                              (plot_contact_matrix, plot_decay_curve,
                               plot_coverage)
  single/hic/_plot.py       → pl/_contact.py + pl/_embedding.py
                              (plot_cell_contacts ↛ pl/_contact.py;
                               plot_embedding ↛ pl/_embedding.py
                               alongside the existing scanpy-style
                               ``embedding`` API — different functions,
                               no namespace clash)

pl/_contact.py is now the single home for every Hi-C heatmap / decay /
coverage helper, regardless of whether the source is a bulk ``.cool``
or a per-cell imputed ``.npz`` from sc-Hi-C.

Backward compatibility:

  - ``epi.bulk.hic.plot_contact_matrix`` etc. remain importable
    (the modality ``__init__.py`` re-imports from ``epione.pl``).
  - ``epi.single.hic.plot_embedding`` and ``plot_cell_contacts`` same.
  - ``epi.hic.plot.X`` deprecation shim updated to point at
    :mod:`epione.pl` rather than the now-removed ``bulk/hic/_plot.py``.
  - Identity verified: ``epi.bulk.hic.plot_contact_matrix is
    epi.pl.plot_contact_matrix`` etc.

Tutorials updated (``t_hic_basics``, ``t_sc_hic_basics``) to prefer
``epi.pl.X`` over the modality-prefixed paths. Both forms continue
to work; new code should use ``epi.pl.X``.

Tests: 71/71 pass.

.gitignore: added ``epione_guide/tutorials/**/result/`` and
``data-tobias-2020/`` so notebook scratch / dataset caches don't get
accidentally committed.
