+++ epione/hic/plot.py
@@ -1,5 +1,5 @@
-"""DEPRECATED: see :mod:`epione.bulk.hic` (canonical location since v0.4)."""
-from epione.bulk.hic._plot import (  # noqa: F401
+"""DEPRECATED: see :mod:`epione.pl` (canonical location since v0.4 PR3)."""
+from epione.pl._contact import (  # noqa: F401
     plot_contact_matrix,
     plot_decay_curve,
     plot_coverage,
